Refactor(#17): 역 정보 가져오는 서비스 위치 변경
