+++ goldmen-application/goldmen-api/src/test/java/com/goldmen/home/user/member/controller/MemberControllerTest.java
@@ -3,5 +3,4 @@ package com.goldmen.home.user.member.controller;
 import static org.junit.jupiter.api.Assertions.*;
 
 class MemberControllerTest {
-
 }
