+++ goldmen-client/seoul-client/src/main/java/com/goldmen/home/station/vo/StationInfoRaw.java
