+++ goldmen-client/seoul-client/src/main/resources/stationInformation.json
